server/requirements.txt
alembic==1.7.1
bcrypt==3.2.0
black==21.9b0
click==8.0.1
cryptography==3.4.8
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.26
pathspec==0.9.0
platformdirs==2.3.0
protobuf==3.17.3
PyMySQL==1.0.2
python-dotenv==0.19.0
regex==2021.8.28
six==1.16.0
SQLAlchemy==1.4.23
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.1
